added owner id, owner specific features also added foreign key and updated schema also increased protection to routes

+++ app/routers/post.py
@@ -3,29 +3,49 @@ from fastapi import FastAPI, HTTPException, status, Response, Depends, APIRouter
 from sqlalchemy.orm import Session
 from ..database import engine, get_db
 from typing import Optional, List
+from . import oauth2
 
 router = APIRouter(
-    prefix="/posts" ,
-    tags =  ["Posts"],
+    prefix="/posts",
+    tags=["Posts"],
 )
 
 
-
-@router.get("/", response_model= List[schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
+@router.get("/", response_model=List[schemas.Post])
+def get_posts(
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+    limit: int = 5,
+    skip: int = 0, search: Optional[str] = "",
+):
     # cursor.execute("""select *from posts""")
     # posts = cursor.fetchall()
     # print(posts)
 
-    posts = db.query(models.Post).all()
+    posts = (
+        db.query(models.Post)
+        .filter(models.Post.owner_id == current_user.id, models.Post.title.contains(search))
+        .limit(limit)
+        .offset(skip)
+        .all()
+    )
+
     return posts
 
 
 # -------/posts path to create post
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-def create_posts(post: schemas.CreatePost, db: Session = Depends(get_db)):
-    post_data = post.model_dump(exclude_unset=True)
+def create_posts(
+    post: schemas.CreatePost,
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+):
+    post_data = post.model_dump(
+        exclude_unset=True
+    )  # Get the dictionary with the existing fields
+    post_data["owner_id"] = current_user.id
     # post_data.pop('rating', None)  # Remove rating if it is not a part of the model
+    print(current_user.email)
     new_post = models.Post(**post_data)
     db.add(new_post)
     db.commit()
@@ -34,20 +54,33 @@ def create_posts(post: schemas.CreatePost, db: Session = Depends(get_db)):
 
 
 # to get a post by id search through my_post via id to get the specific post
-@router.get("/{post_id}" , response_model= schemas.Post)
-def get_post(post_id: int, db: Session = Depends(get_db)):
+@router.get("/{post_id}", response_model=schemas.Post)
+def get_post(
+    post_id: int,
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+):
     post = db.query(models.Post).filter(models.Post.id == post_id).first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with id: {post_id} is not found!",
         )
+    if post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to perform requested action",
+        )
     return post
 
 
 # Deleting the post-----
 @router.delete("/delete/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(post_id: int, db: Session = Depends(get_db)):
+def delete_post(
+    post_id: int,
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+):
     delete_post = db.query(models.Post).filter(models.Post.id == post_id).first()
 
     if delete_post == None:
@@ -55,20 +88,37 @@ def delete_post(post_id: int, db: Session = Depends(get_db)):
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with id {post_id} not found!",
         )
+
+    if delete_post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to perform requested action",
+        )
     db.delete(delete_post)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
 # ---Update the post
-@router.put("/{post_id}",  response_model= schemas.Post)
-def update_post(post_id: int, post: schemas.CreatePost, db: Session = Depends(get_db)):
+@router.put("/{post_id}", response_model=schemas.Post)
+def update_post(
+    post_id: int,
+    post: schemas.CreatePost,
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+):
     updated_post = db.query(models.Post).filter(models.Post.id == post_id).first()
 
     if updated_post is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with id: {post_id} does not exist!",
+        )
+
+    if updated_post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to perform requested action",
         )
     for key, value in post.model_dump(exclude_unset=True).items():
         setattr(updated_post, key, value)
